Chnages to IPAT
git-svn-id: https://svn.code.sf.net/p/directshownet/code/trunk@973 b2a09c84-0b36-484d-b36e-ffab392acda3

+++ directshowlib/src/mpeg2psiparser.cs
@@ -88,7 +88,7 @@ namespace DirectShowLib.BDA
     int GetVersionNumber([Out] out byte pbVal);
 
     [PreserveSig]
-    int GetCountOfRecords([Out] out short pwVal);
+    int GetCountOfRecords([Out] out int pwVal);
 
     [PreserveSig]
     int GetRecordProgramNumber(
